Try applying to subscribe  before resolve
If an explicit resolver exists on a subscription the middleware is
being applied to it instead of the subscribe function.  This change
checks for subscribe and applies middleware on it first, which seems
more useful and consistent (for example, in graphql-shield).

In *theory* a user of this library could be relying on this behavior to
wrap resolve instead of subscribe, but given it only does it with
explicitly defined resolvers it would not be handling that well anyway.